Prefer zstd compression in .7z files
zstd is much faster to uncompress than 7za's default lzma2.

https://jira.unity3d.com/browse/UBS-316

+++ external/buildscripts/collect_allbuilds.pl
@@ -11,6 +11,11 @@ my $monoroot = abs_path($monoroot);
 
 my $path = "incomingbuilds/";
 
+if ($^O ne "linux" && $^O ne "darwin")
+{
+	die("Unsupported platform to create stevedore artifact.")
+}
+
 rmtree("collectedbuilds");
 mkpath("collectedbuilds");
 
@@ -57,72 +62,32 @@ system("cp embedruntimes/linux64/libMonoPosixHelper.so monodistribution/lib/libM
 
 rmove('versions-aggregated.txt', 'versions.txt');
 
-my $externalBuildDeps = "$monoroot/external/buildscripts/artifacts/Stevedore";
-my $externalzip = "";
-if($^O eq "linux")
-{
-	$externalzip = "$externalBuildDeps/7za-linux-x64/7za";
-}
-elsif($^O eq 'darwin')
-{
-	$externalzip = "$externalBuildDeps/7za-mac-x64/7za";
-}
-
-
-# Create stevedore artifact
 print(">>> Create stevedore artifact $monoroot/stevedore/MonoBleedingEdge.7z");
-if($^O eq "linux" || $^O eq 'darwin')
-{
-	rmtree("../stevedore");
-	my $stevedoreMbePath = "../stevedore/MonoBleedingEdge";
-	my $stevedoreMbe7z = "../stevedore/MonoBleedingEdge.7z";
-	my $stevedoreMbeArtifactID = "../stevedore/artifactid.txt";
 
-	system("mkdir -p $stevedoreMbePath") eq 0 or die("failed to mkdir $stevedoreMbePath");
-	system("cp -r * $stevedoreMbePath/") eq 0 or die ("failed copying builds to $stevedoreMbePath\n");
-	if(-f $externalzip)
-	{
-		system("$externalzip a $stevedoreMbe7z $stevedoreMbePath/* -sdel") eq 0 or die("failed 7z up $stevedoreMbePath");
-	}
-	else
-	{
-		#Use 7z installed on the machine. If its not installed, please install it.
-		system("7z a $stevedoreMbe7z $stevedoreMbePath/* -sdel") eq 0 or die("failed 7z up $stevedoreMbePath");
-	}
-	system("rm -rf $stevedoreMbePath") eq 0 or die("failed to delete $stevedoreMbePath");
+rmtree("../stevedore");
+my $stevedoreMbePath = "../stevedore/MonoBleedingEdge";
+my $stevedoreMbe7z = "../stevedore/MonoBleedingEdge.7z";
+my $stevedoreMbeArtifactID = "../stevedore/artifactid.txt";
 
-	# Write stevedore artifact ID to file
-	my $revision = `git rev-parse --short HEAD`;
-	system("mono ../external/buildscripts/bee.exe steve new $stevedoreMbe7z MonoBleedingEdge $revision") eq 0 or die("failed running bee");
-	open (my $file, '>', $stevedoreMbeArtifactID);
-	my $artifactID = `mono ../external/buildscripts/bee.exe steve new $stevedoreMbe7z MonoBleedingEdge $revision`;
-	print $file $artifactID; 
-	print (">>> MonoBleedingEdge stevedore artifact ID: $artifactID\n");
-}
-else
-{
-	die("Unsupported platform to create stevedore artifact.")
-}
+system("mkdir -p $stevedoreMbePath") eq 0 or die("failed to mkdir $stevedoreMbePath");
+system("cp -r * $stevedoreMbePath/") eq 0 or die ("failed copying builds to $stevedoreMbePath\n");
+
+# Use our 7za fork with zstd support, installed in the PATH somewhere.
+# See .yamato/scripts/collate_builds.sh for more details.
+system("7za a $stevedoreMbe7z -m0=zstd -mx22 $stevedoreMbePath/* -sdel") eq 0 or die("failed 7z up $stevedoreMbePath");
+system("rm -rf $stevedoreMbePath") eq 0 or die("failed to delete $stevedoreMbePath");
+
+# Write stevedore artifact ID to file
+my $revision = `git rev-parse --short HEAD`;
+system("mono ../external/buildscripts/bee.exe steve new $stevedoreMbe7z MonoBleedingEdge $revision") eq 0 or die("failed running bee");
+open (my $file, '>', $stevedoreMbeArtifactID);
+my $artifactID = `mono ../external/buildscripts/bee.exe steve new $stevedoreMbe7z MonoBleedingEdge $revision`;
+print $file $artifactID; 
+print (">>> MonoBleedingEdge stevedore artifact ID: $artifactID\n");
+
 print(">>> Done creating stevedore artifact $monoroot/MonoBleedingEdge.7z");
 
 system("zip -r builds.zip *") eq 0 or die("failed zipping up builds");
 
-if($^O eq "linux" || $^O eq 'darwin')
-{
-	if(-f $externalzip)
-	{
-		system("$externalzip a builds.7z * -x!builds.zip") eq 0 or die("failed 7z up builds");
-	}
-	else
-	{
-		#Use 7z installed on the machine. If its not installed, please install it.
-		system("7z a builds.7z * -x!builds.zip") eq 0 or die("failed 7z up builds");
-	}
-}
-else
-{
-	die("Unsupported platform for build collection.")
-}
-
-
+system("7za a builds.7z -m0=zstd -mx22 * -x!builds.zip") eq 0 or die("failed 7z up builds");
